Add new internationalization (i18n) file
Introduced an English translation file (en.pot) under the shared packages in i18n. This file contains translated string messages that serve as a basis for software localization. All phrases displayed to the user, and their translations, are now stored here. This allows for easier translation updates and ensures consistency in terminology across the application.